chore: move status message to shared folder
